requirements.txt
aniso8601==9.0.1
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.1
click==8.0.1
colorama==0.4.4
cssselect==1.1.0
dnspython==1.16.0
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.0.12
Flask==2.0.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
idna==3.2
itsdangerous==2.0.1
jieba3k==0.35.1
Jinja2==3.0.1
joblib==1.0.1
lxml==4.6.3
MarkupSafe==2.0.1
newspaper3k==0.2.8
nltk==3.6.2
Pillow==8.3.1
pymongo==3.12.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.7.6
requests==2.26.0
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.2.1
tinysegmenter==0.3
tldextract==3.1.0
tqdm==4.61.2
urllib3==1.26.6
Werkzeug==2.0.1
